requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.8
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
spotipy==2.25.0
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
